Feature: Add new command (Smart Change Directory)
This feature introduces scd, a command that changes directory
according to a `.smartcdfolders` file. This way I can change
from anywhere to some folders that is special. The usage is simple

	. scd foldername

The folder to where we want to cd must be inside one of the directory
listed on the special file.